# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Browser4 WebDriver-Compatible API
  description: |
    A WebDriver-compatible HTTP API with selector-first extensions for browser automation.
    This API provides session management, navigation, element interaction, and event handling
    capabilities.
  version: 1.0.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8182
    description: Local development server
...
paths:
...
  /session/{sessionId}/control/delay:
    post:
      tags:
        - control
      summary: Delay execution
      operationId: delay
      parameters:
        - $ref: '#/components/parameters/sessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DelayRequest'
      responses:
        '200':
          description: Delay completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebDriverResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/control/pause:
    post:
      tags:
        - control
      summary: Pause session execution
      operationId: pause
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: Session paused
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebDriverResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/control/stop:
    post:
      tags:
        - control
      summary: Stop session execution
      operationId: stop
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: Session stopped
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebDriverResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  parameters:
    sessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
      description: Session identifier
...
  schemas:
...
    WebDriverResponse:
      type: object
      properties:
        value:
          description: Response value (can be null for success)
    ErrorResponse:
      type: object
      properties:
        value:
          type: object
          properties:
            error:
              type: string
            message:
              type: string
            stacktrace:
              type: string
...
    DelayRequest:
      type: object
      required:
        - ms
      properties:
        ms:
          type: integer
          minimum: 0
          description: Delay duration in milliseconds
...